requirements.txt
azure-functions
openai>=1.0.0
python-dotenv
pdfplumber
pandas
numpy
azure-cosmos
tiktoken